+++ CollectorDemoApp/CollectorDemoAppCocoaPods/MainController.swift
@@ -1,5 +1,0 @@
-import Foundation
-import UIKit
-import BitmovinAnalyticsCollector
-
-class MainViewController: UIViewController {}