+++ Package.swift
@@ -6,8 +6,8 @@ import PackageDescription
 let package = Package(
     name: "BitmovinAnalytics",
     platforms: [
-        .iOS(.v12),
-        .tvOS(.v12)
+        .iOS("14.2"),
+        .tvOS("14.2")
     ],
     products: [
         // Products define the executables and libraries a package produces, and make them visible to other packages.